requirements.txt
chromadb
fastapi
groq
PyMuPDF
python-dotenv
python-multipart
uvicorn
